fix: add support for variable-length array DIEs
:subrange_type DIEs with a count set to a :variable DIE would raise an
exception in Die#type_name.  This was because we weren't properly
supporting variable-length arrays declared in the .debug_info.  These
can be created on the stack in c using:

```c
void blah(unsigned count) {
    unsigned my_array[count];
}
```

In this commit I add support for these DIEs, along with tests to verify
they work properly across all supported versions.  There is one
important note here, the CTypes array returned for these DIEs are
greedy.  This is because there is no way to know how many elements are
present at run-time.