Die#ctype: support tail-padded structs
We weren't accounting for the fact that the DW_AT_byte_size may be
larger than the sum of DW_TAG_member type sizes.  As a result, the ctype
structs generated were smaller than they should be, causing problems
when working with arrays of tail-padded structs.

In this commit we add the necessary padding if the struct is smaller
than the DW_AT_byte_size defined in the DW_TAG_structure_type DIE.

fixes #3